+++ R/polar_freqMap.R
@@ -84,6 +84,7 @@ freqMap <- function(data,
                     d.fig = 3.5,
                     static = FALSE,
                     static.nrow = NULL,
+                    progress = TRUE,
                     ...,
                     control = NULL) {
   # check basemap providers are valid
@@ -214,7 +215,8 @@ freqMap <- function(data,
       split_col = split_col,
       d.fig = d.fig,
       popup = popup,
-      label = label
+      label = label,
+      progress = progress
     )
 
   if (static) {