+++ R/polar_windroseMap.R
@@ -77,6 +77,7 @@ windroseMap <- function(data,
                         d.fig = 3.5,
                         static = FALSE,
                         static.nrow = NULL,
+                        progress = TRUE,
                         ...,
                         control = NULL) {
   # check basemap providers are valid
@@ -173,7 +174,8 @@ windroseMap <- function(data,
       split_col = split_col,
       d.fig = d.fig,
       popup = popup,
-      label = label
+      label = label,
+      progress = progress
     )
 
   if (static) {