+++ NEWS.md
@@ -1,3 +1,5 @@
+# openairmaps (development version)
+
 # openairmaps 0.9.0
 
 ## Breaking changes