+++ NEWS.md
@@ -12,6 +12,8 @@
 
 * `...` will successfully pass to `openair::cutData()` in the `polarMap()` and `trajMap()` families of functions.
 
+* The default `diffMap()` colour scale will no longer appear inverted compared to `openair::polarDiff()`.
+
 # openairmaps 0.9.0
 
 ## Breaking changes