+++ NEWS.md
@@ -1,4 +1,4 @@
-# openairmaps (development version)
+# openairmaps 0.9.1
 
 ## New features